Removed error page redirection and minor changes
If clear or process button is pressed but no image is found, doesn't crash.

If upload button is pressed but no image is selected, doesn't crash.

Changed the max size of the image to be 500.

Made it so when the user uploads a new image it deletes the old one.