add UUID to upload and begin fixing cleanup
The app now creates a folder from a UUID for each user (session), and this is the upload folder for them.

Also added a timeout timer, so if they are inactive for 30 mins the directory is deleted. It does not work if they close the browser, but it is a step in the right direction.